change file name from bankroll page to bankroll header, create new file for bankroll page
