+++ apps/poker-roll/src/app/bankrollHeader.jsx
@@ -1,0 +1,21 @@
+import React from "react";
+import Card from '@mui/material/Card';
+import CardContent from '@mui/material/CardContent';
+import Typography from '@mui/material/Typography';
+import { Button, CardActionArea, CardActions } from '@mui/material';
+
+const BankrollHeader = (props) => {
+
+    return (
+        <div>
+            <Card sx={{ maxWidth: 345 }}>
+                    <CardContent>
+                        <Typography gutterBottom variant="h5" component="div">
+                            Current BankRoll: $0.00
+                        </Typography>
+                    </CardContent>    
+            </Card>
+        </div>
+    )
+}
+export default BankrollHeader;
